big refactoring
Collapse WasmBase, ContextBase, PluginBase and rest of all the **Base
remove source/host
