+++ contrib/http_wasm/filters/http/source/http_wasm_common.h
