+++ contrib/http_wasm/filters/http/source/v8/v8.h
@@ -2,14 +2,16 @@
 
 #include <memory>
 
-#include "contrib/http_wasm/filters/http/source/host/vm_runtime.h"
+#include "contrib/http_wasm/filters/http/source/vm_runtime.h"
 namespace Envoy {
 namespace Extensions {
 namespace HttpFilters {
 namespace HttpWasm {
+namespace V8 {
 
-std::unique_ptr<Host::WasmVm> createV8Vm();
+std::unique_ptr<HttpWasm::Runtime> createV8Vm();
 
+}
 } // namespace HttpWasm
 } // namespace HttpFilters
 } // namespace Extensions
